Filter rangeland layer
